+++ src/mechanics/JewelMechanics.h
@@ -47,7 +47,7 @@ public:
     ~JewelMechanics();
     bool getCoinFlip(double) const;
     unsigned int calculateStat(const Pokemon &p, const STAT i) const;
-    int calculateDamage(BattleField &field, MoveObject &move,
+    unsigned long calculateDamage(BattleField &field, MoveObject &move,
             Pokemon &user, Pokemon &target, const int targets,
             const bool weight = true) const;
     bool isCriticalHit(BattleField &field, MoveObject &move,
